modify data type

+++ data_process/build_dataloader.py
@@ -75,9 +75,9 @@ class ImgTxtDataSet(Dataset):
 
 
 def batch_dataset_fn(batch):
-    batch_patch = torch.tensor([b[0] for b in batch])
-    batch_tokens = torch.tensor([b[1] for b in batch], dtype=torch.uint8)
-    batch_labels = torch.tensor([b[2] for b in batch], dtype=torch.uint8)
+    batch_patch = torch.tensor([b[0] for b in batch], dtype=torch.float32)
+    batch_tokens = torch.tensor([b[1] for b in batch], dtype=torch.int32)
+    batch_labels = torch.tensor([b[2] for b in batch], dtype=torch.int32)
     return batch_patch, batch_tokens, batch_labels
